Add some Floppy Disc Controler function
Add "send command" and "read result" used to send and recieve all the command to he FDD controler, the 2 function test if the FDC is readdy to resive/send the data, then sent the right number of data.
I aldo addes the fonction recalibrate to bring back the head back to the track 0.